% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nass_set_key.R
\name{nass_set_key}
\alias{nass_set_key}
\title{Install a NASS Quickstats API Key in Your \code{.Renviron} File for
  Repeated Use}
\usage{
nass_set_key(token = NA, overwrite = NA)
}
\arguments{
\item{token}{The API key provided to you from NASS formated in quotes. A key
can be acquired at \url{https://quickstats.nass.usda.gov/api}}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing
NASS_KEY that you already have in your \code{.Renviron} file.}
}
\description{
This function will add your NASS Quickstats API key to your
  \code{.Renviron} file so it can be called securely without being stored in
  your code. After you have installed your key, it can be called any time by
  typing \code{Sys.getenv("NASS_KEY")} and can be used in package functions
  by simply typing NASS_KEY. If you do not have an \code{.Renviron} file, the
  function will create on for you. If you already have an \code{.Renviron}
  file, the function will append the key to your existing file, while making
  a backup of your original file for disaster recovery purposes.
}
\examples{

\donttest{
set_nass_key("111111abc")
# First time, relead your enviornment so you can use the key without
# restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("NASS_KEY")
}

\donttest{
# If you need to overwrite an existing key:
nass_set_key("111111abc", overwrite = TRUE)
# First time, relead your enviornment so you can use the key without
# restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("NASS_KEY")
}
}
